zulint: Move pycodestyle runner code to library.
This was, only our configuration ends up living in pep8.py.
